requirements.txt
# Marketplace Unidades Residenciales
# Python 3.10+

django>=4.2
djangorestframework>=3.14

# Opcional para validaciones y configuración futura
# pydantic>=2.0
# python-dotenv>=1.0
